WIP: Double check this!
Majorly improved memory usage by replacing the Iteration struct with usize, and making `escape_time()` returned the smooth palette index instead of the iteration.

It's a bit 'incorrect' mathematically, but it means we only have to store usize's in the pixels array, instead og whole structs with Bignums in them.